Run 'autoconf'
Whilst it's not-done to keep auto-generated files in a VCS, this
repository is an exception on this rule for manageability reasons.

This commit updates the 'configure' script according to the changes
made to 'configure.in' in 14fe126eb8cb7af36e97ca023f258c109bcbde97.

See: 14fe126eb8cb7af36e97ca023f258c109bcbde97